pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "git-filter-repo>=2.47.0",
    "ipython>=9.10.0",
    "jupter>=0.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "pandas>=3.0.0",
    "peft>=0.18.1",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.1.0",
]
